Add layer geometry in hazard.
