+++ safe/gui/tools/wizard/step_kw00_purpose.py
@@ -29,7 +29,7 @@ from safe.definitionsv4.utilities import purposes_for_layer
 from safe.gui.tools.wizard.wizard_step import WizardStep
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_strings import category_question
-from safe.utilities.keyword_io import definition
+from safe.definitionsv4.utilities import definition
 from safe.utilities.resources import resources_path
 
 FORM_CLASS = get_wizard_step_ui_class(__file__)
