+++ safe/gui/tools/wizard/step_kw05_subcategory.py
@@ -29,7 +29,7 @@ from safe.definitionsv4.layer_purposes import (
 from safe.gui.tools.wizard.wizard_step import WizardStep
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_utils import get_question_text
-from safe.utilities.keyword_io import definition
+from safe.definitionsv4.utilities import definition
 from safe.utilities.resources import resources_path
