+++ safe/gui/tools/wizard/step_kw10_hazard_category.py
@@ -27,7 +27,7 @@ from safe.definitionsv4.layer_purposes import layer_purpose_hazard
 from safe.gui.tools.wizard.wizard_step import WizardStep
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_strings import hazard_category_question
-from safe.utilities.keyword_io import definition
+from safe.definitionsv4.utilities import definition
 
 FORM_CLASS = get_wizard_step_ui_class(__file__)
