+++ safe/gui/tools/wizard/step_kw25_classification.py
@@ -27,7 +27,7 @@ from safe.gui.tools.wizard.wizard_step import WizardStep
 from safe.gui.tools.wizard.wizard_step import get_wizard_step_ui_class
 from safe.gui.tools.wizard.wizard_strings import classification_question
 from safe.utilities.gis import is_raster_layer
-from safe.utilities.keyword_io import definition
+from safe.definitionsv4.utilities import definition
 
 FORM_CLASS = get_wizard_step_ui_class(__file__)
